chore(*): bump golangci-lint (#1799)
Bump golangci-lint from v1.50.1 to v1.60.3

Add `ToInt*` int conversion functions to `umath` package.

issue: none

---------

+++ e2e/vmcompose/data.go
@@ -16,8 +16,8 @@ import (
 var omniEvmRegx = regexp.MustCompile(".*_evm")
 
 const (
-	evmPort  = 8545
-	haloPort = 26657
+	evmPort  = uint32(8545)
+	haloPort = uint32(26657)
 	relayer  = "relayer"
 )
 
@@ -76,7 +76,7 @@ func LoadData(path string) (types.InfrastructureData, error) {
 		instances[serviceName] = e2e.InstanceData{
 			IPAddress:    vm.IPAddress,
 			ExtIPAddress: vm.ExtIPAddress,
-			Port:         uint32(port),
+			Port:         port,
 		}
 	}
